Fix crash on reg dialog (old bug shows up due to more restrict wxWidgets 3.1)
main fix is a crash in regression dialog on windows:  it is caused by 1.
select variable to independent variable list and 2. remove the
independent variable. The program will prompt error message dialog,
because the code is trying to activate not-existing item in the
wxListbox.

This bug has been there all the time (back to 1.6.x). It has been
activated because of the upgrade of wxWidgets to version 3.1, which is
very restrict on operations of wxListBox.

Other updates are for eliminating the compile warning on Windows.

+++ Explore/ConnectivityMapView.cpp
@@ -288,7 +288,7 @@ void ConnectivityMapCanvas::UpdateFromSharedCore()
 {
 	sel_cores.clear();
 	std::vector<bool>& sc_hs = shared_core_hs->GetHighlight();
-	for (size_t i=0, sz=sc_hs.size(); i<sz; ++i) {
+	for (int i=0, sz=sc_hs.size(); i<sz; ++i) {
 		if (sc_hs[i]) sel_cores.insert(i);
 	}
